refactor(core): Wave 5 — final 3 borderline files with domain-driven splits
- pq.rs (513→299): Repository pattern → pq_persistence.rs (59)
  Persistence (postcard serde, atomic file I/O) separated from PQ algorithm
- perf_optimizations.rs (514→178): Extract Class → contiguous_resize.rs (56)
  Resize/realloc memory management separated from storage API
- traversal.rs (501→297): Composition Layer → traversal_bidir.rs (32)
  Bidirectional BFS composes unidirectional BFS, no access to internals

All production files now under 500 NLOC. Zero public API changes.
