+++ crates/velesdb-core/src/collection/graph/traversal.rs
@@ -449,53 +449,4 @@ pub fn bfs_traverse_reverse(
     bfs_traverse_directed(edge_store, source_id, config, BfsDirection::Reverse)
 }
 
-/// Performs bidirectional BFS (follows both directions).
-///
-/// # Deduplication strategy
-///
-/// Uses target-only dedup via `FxHashSet<u64>`: each target node appears
-/// at most once across forward and reverse results. This was chosen over
-/// the prior path+target dedup because:
-/// - O(1) `HashSet::contains` vs O(n) linear scan per reverse result.
-/// - Cleaner output — a node reached by both forward and reverse edges
-///   appears once (via whichever direction discovered it first).
-/// - The forward pass populates the `seen` set; reverse results are skipped
-///   if their `target_id` is already present.
-#[must_use]
-pub fn bfs_traverse_both(
-    edge_store: &EdgeStore,
-    source_id: u64,
-    config: &TraversalConfig,
-) -> Vec<TraversalResult> {
-    let mut results = Vec::new();
-    let half_limit = config.limit / 2 + 1;
-
-    let config_half = TraversalConfig {
-        limit: half_limit,
-        ..config.clone()
-    };
-
-    // Forward traversal
-    let forward = bfs_traverse(edge_store, source_id, &config_half);
-    // Build O(1) dedup set from forward results to avoid O(n) linear scan per reverse result.
-    let seen: FxHashSet<u64> = forward.iter().map(|r| r.target_id).collect();
-    results.extend(forward);
-
-    // Reverse traversal — skip targets already reached by forward BFS
-    if results.len() < config.limit {
-        let reverse = bfs_traverse_reverse(edge_store, source_id, &config_half);
-        for r in reverse {
-            if results.len() >= config.limit {
-                break;
-            }
-            if !seen.contains(&r.target_id) {
-                results.push(r);
-            }
-        }
-    }
-
-    results.truncate(config.limit);
-    results
-}
-
 // Tests moved to traversal_tests.rs per project rules
